re-org extensions
